requirements.txt
faiss-cpu>=1.7.4
numpy>=1.24.0
prompt_toolkit>=3.0.39
Pygments>=2.15.0
sentence_transformers>=2.2.2
setuptools>=65.0.0
torch>=2.0.0
